+++ findBaseAddress.cpp
@@ -70,50 +70,3 @@ void ProcessManager::FindBaseAddress(char* Module)
 	cout <<"Span = "<<span << endl;//======================================================================================
 	free(FileBuffer);
 }
-
-void ProcessManager::PrintContent()
-{
-	struct iovec *local = new struct iovec;
-	char* op = new char[span];
-	local->iov_base = op;
-	local->iov_len = span;
-	struct iovec *remote =new struct iovec;
-	remote->iov_base =(void*)BaseAddress;
-	remote->iov_len = span;
-	ssize_t nread;
-	nread = process_vm_readv(pID, local, 1, remote, 1, 0);
-	for (int i = 0;i<nread;i++)
-	{
-		cout <<(void*)(BaseAddress+i)<<": "<< hex << (int)op[i]<<endl;
-	}
-	if (nread != span)
-		cout << "Full Data Read failed due to unaccessibility of a memory location\n";
-	free(op);
-}
-
-int main()
-{
-	long id,BA;
-	cout << "Enter PID of the Process\n";
-	cin >> id;
-	ProcessManager P(id);
-	P.span = 16;
-	// P.FindBaseAddress("[stack]");
-	// P.PrintContent();
-	// cout << "Enter BaseAddress\n";
-	// cin >> BA;
-	// ProcessManager P(id,BA,16);
-
-	// P.getContent("int");
-
-	char test[16] = "7f4644c29001";
-	// int **add;
-	P.getAddresses();
-	// int size ;
-	// // size = P.AddressesRows;
-	// cout << "span:" << P.span << endl;
-	// cout << "Addresses size: "<< P.AddressesRows <<endl;
-	P.identifierFunction(test);
-	// cout <<  P.Addresses  << endl << add <<endl;
-	return 0;
-}
